Updated Bottle Synonyms and Door Locks
Added essence bottle synonyms. The bathroom door and kitchen door are now locked at the start of the game. Sulfuric acid is needed to open the bathroom door while the kitchen key is needed to open the kitchen door. For testing purposes, alter setLocked() from true to false in the constructor to open the doors.